+++ packages/client/src/views/VerifyCertificate/useVerificationRecordDetails.ts
@@ -49,12 +49,16 @@ const FETCH_RECORD_DETAILS_FOR_VERIFICATION = gql`
           action
           regStatus
           user {
+            primaryOffice {
+              hierarchy {
+                id
+                name
+                alias
+              }
+            }
             name {
               firstNames
               familyName
-            }
-            catchmentArea {
-              name
             }
           }
         }
@@ -94,12 +98,16 @@ const FETCH_RECORD_DETAILS_FOR_VERIFICATION = gql`
           action
           regStatus
           user {
+            primaryOffice {
+              hierarchy {
+                id
+                name
+                alias
+              }
+            }
             name {
               firstNames
               familyName
-            }
-            catchmentArea {
-              name
             }
           }
         }