+++ packages/config/src/config/routes.ts
@@ -35,7 +35,7 @@ import getForms from '@config/handlers/forms/formsHandler'
 import getDashboardQueries from '@config/handlers/dashboardQueries/dashboardQueries'
 import { ServerRoute } from '@hapi/hapi'
 import * as Joi from 'joi'
-import { resolveChildren } from '@config/handlers/locations/resolveChildren'
+import { resolveChildren } from '@config/handlers/locations/children'
 import {
   fetchLocationsHandler,
   locationQuerySchema,
@@ -269,21 +269,7 @@ export default function getRoutes(): ServerRoute[] {
         }
       }
     },
-    {
-      method: 'GET',
-      path: '/location/{locationId}/hierarchy',
-      handler: locationHierarchyHandler,
-      options: {
-        tags: ['api'],
-        auth: false,
-        description: "Get location's hierarchy",
-        validate: {
-          params: Joi.object({
-            locationId: Joi.string().uuid()
-          })
-        }
-      }
-    },
+
     {
       method: 'GET',
       path: '/dashboardQueries',
@@ -307,19 +293,6 @@ export default function getRoutes(): ServerRoute[] {
       }
     },
     {
-      method: 'GET',
-      path: '/locations/{locationId}',
-      handler: fetchLocationHandler,
-      options: {
-        tags: ['api'],
-        auth: false,
-        description: 'Get a single location',
-        validate: {
-          params: requestParamsSchema
-        }
-      }
-    },
-    {
       method: 'POST',
       path: '/locations',
       handler: createLocationHandler,
@@ -331,6 +304,19 @@ export default function getRoutes(): ServerRoute[] {
         description: 'Create a location',
         validate: {
           payload: createLocationReqSchema
+        }
+      }
+    },
+    {
+      method: 'GET',
+      path: '/locations/{locationId}',
+      handler: fetchLocationHandler,
+      options: {
+        tags: ['api'],
+        auth: false,
+        description: 'Get a single location',
+        validate: {
+          params: requestParamsSchema
         }
       }
     },
@@ -347,6 +333,21 @@ export default function getRoutes(): ServerRoute[] {
         validate: {
           payload: updateSchema,
           params: requestParamsSchema
+        }
+      }
+    },
+    {
+      method: 'GET',
+      path: '/locations/{locationId}/hierarchy',
+      handler: locationHierarchyHandler,
+      options: {
+        tags: ['api'],
+        auth: false,
+        description: "Get location's hierarchy",
+        validate: {
+          params: Joi.object({
+            locationId: Joi.string().uuid()
+          })
         }
       }
     },