+++ packages/config/src/handlers/locations/hierarchy.ts
@@ -12,9 +12,17 @@ import { ServerRoute } from '@hapi/hapi'
 import { fetchLocations } from '@config/services/hearth'
 import { resolveLocationParents } from './locationTreeSolver'
 import { UUID } from '@opencrvs/commons'
+import { find } from 'lodash'
+import { notFound } from '@hapi/boom'
 
 export const locationHierarchyHandler: ServerRoute['handler'] = async (req) => {
   const { locationId } = req.params as { locationId: UUID }
   const locations = await fetchLocations()
-  return resolveLocationParents(locationId, locations)
+  const location = find(locations, { id: locationId })
+
+  if (!location) {
+    return notFound()
+  }
+
+  return [...resolveLocationParents(location, locations), location]
 }