+++ packages/gateway/src/features/registration/type-resolvers.ts
@@ -55,6 +55,7 @@ import {
   INFORMANT_CODE,
   Identifier,
   Location,
+  SavedLocation,
   MOTHER_CODE,
   OPENCRVS_SPECIFICATION_URL,
   Patient,
@@ -82,7 +83,9 @@ import {
   isQuestionnaireResponse,
   isTaskOrTaskHistory,
   resourceIdentifierToUUID,
-  Address
+  Address,
+  findLastOfficeFromSavedBundle,
+  findLastOfficeLocationFromSavedBundle
 } from '@opencrvs/commons/types'
 
 import { GQLQuestionnaireQuestion, GQLResolver } from '@gateway/graphql/schema'
@@ -1094,10 +1097,8 @@ export const typeResolvers: GQLResolver = {
         `${OPENCRVS_SPECIFICATION_URL}extension/regAssigned`,
         task.extension
       )
-      const regLastOfficeExtension = findExtension(
-        `${OPENCRVS_SPECIFICATION_URL}extension/regLastOffice`,
-        task.extension
-      )
+      const office =
+        context.record && findLastOfficeFromSavedBundle(context.record)
 
       if (assignmentExtension) {
         const regLastUserExtension = findExtension(
@@ -1113,12 +1114,13 @@ export const typeResolvers: GQLResolver = {
             await context.dataSources.usersAPI.getUserByPractitionerId(
               practitionerId
             )
+
           if (user) {
             return {
               userId: user._id,
               firstName: user.name[0].given.join(' '),
               lastName: user.name[0].family,
-              officeName: regLastOfficeExtension?.valueString || ''
+              officeName: office?.name || ''
             }
           }
         }
@@ -1145,19 +1147,15 @@ export const typeResolvers: GQLResolver = {
     reason: (task: Task) => (task.reason && task.reason.text) || null,
     timestamp: (task) => task.lastModified,
     comments: (task) => task.note,
-    location: async (task, _, { dataSources }) => {
+    location: async (task, _, { dataSources, record }) => {
       const taskLocation = findExtension(
-        `${OPENCRVS_SPECIFICATION_URL}extension/regLastLocation`,
-        task.extension
+        `${OPENCRVS_SPECIFICATION_URL}extension/regLastOffice`,
+        task.extension as Extension[]
       )
-      if (!taskLocation || !taskLocation.valueReference) {
+      if (!taskLocation || !record) {
         return null
       }
-      const found = await dataSources.locationsAPI.getLocation(
-        taskLocation.valueReference.reference?.split('/')[1] as string
-      )
-
-      return found
+      return findLastOfficeLocationFromSavedBundle(record)
     },
     office: async (task, _, { dataSources }) => {
       const taskLocation = findExtension(
@@ -1168,7 +1166,7 @@ export const typeResolvers: GQLResolver = {
         return null
       }
       return dataSources.locationsAPI.getLocation(
-        taskLocation.valueReference.reference?.split('/')[1] as string
+        resourceIdentifierToUUID(taskLocation.valueReference.reference)
       )
     },
     timeLogged: async (task, _, { dataSources }) => {
@@ -1304,7 +1302,10 @@ export const typeResolvers: GQLResolver = {
         null
       )
     },
-    address: (location) => location.address
+    address: (location: Location) => location.address,
+    hierarchy: (location: SavedLocation, _, { dataSources }) => {
+      return dataSources.locationsAPI.getHierarchy(location.id)
+    }
   },
   MedicalPractitioner: {
     name: async (encounterParticipant: EncounterParticipant, _, context) => {
@@ -1566,17 +1567,15 @@ export const typeResolvers: GQLResolver = {
       }
       return JSON.parse(systemIdentifier.value)
     },
-    location: async (task: Task, _: any, { dataSources }) => {
+    location: async (task: Task, _: any, { dataSources, record }) => {
       const taskLocation = findExtension(
-        `${OPENCRVS_SPECIFICATION_URL}extension/regLastLocation`,
+        `${OPENCRVS_SPECIFICATION_URL}extension/regLastOffice`,
         task.extension as Extension[]
       )
-      if (!taskLocation || !taskLocation.valueReference) {
+      if (!taskLocation || !record) {
         return null
       }
-      return dataSources.locationsAPI.getLocation(
-        taskLocation.valueReference.reference?.split('/')[1] as string
-      )
+      return findLastOfficeLocationFromSavedBundle(record)
     },
     office: async (task: Task, _: any, { dataSources }) => {
       const taskLocation = findExtension(