+++ packages/gateway/src/graphql/schema.d.ts
@@ -289,6 +289,7 @@ export interface GQLLocation {
   latitude?: number
   altitude?: number
   geoData?: string
+  hierarchy?: Array<GQLLocation>
 }
 
 export interface GQLUser {
@@ -304,7 +305,6 @@ export interface GQLUser {
   status: GQLStatus
   underInvestigation?: boolean
   primaryOffice?: GQLLocation
-  catchmentArea?: Array<GQLLocation>
   localRegistrar?: GQLLocalRegistrar
   identifier?: GQLIdentifier
   signature?: GQLSignature
@@ -674,7 +674,6 @@ export interface GQLUserInput {
   role?: string
   email?: string
   primaryOffice?: string
-  catchmentArea?: Array<string | null>
   device?: string
   signature?: GQLSignatureInput
 }
@@ -4854,6 +4853,7 @@ export interface GQLLocationTypeResolver<TParent = any> {
   latitude?: LocationToLatitudeResolver<TParent>
   altitude?: LocationToAltitudeResolver<TParent>
   geoData?: LocationToGeoDataResolver<TParent>
+  hierarchy?: LocationToHierarchyResolver<TParent>
 }
 
 export interface LocationToIdResolver<TParent = any, TResult = any> {
@@ -4991,6 +4991,15 @@ export interface LocationToGeoDataResolver<TParent = any, TResult = any> {
   ): TResult
 }
 
+export interface LocationToHierarchyResolver<TParent = any, TResult = any> {
+  (
+    parent: TParent,
+    args: {},
+    context: Context,
+    info: GraphQLResolveInfo
+  ): TResult
+}
+
 export interface GQLUserTypeResolver<TParent = any> {
   id?: UserToIdResolver<TParent>
   userMgntUserID?: UserToUserMgntUserIDResolver<TParent>
@@ -5004,7 +5013,6 @@ export interface GQLUserTypeResolver<TParent = any> {
   status?: UserToStatusResolver<TParent>
   underInvestigation?: UserToUnderInvestigationResolver<TParent>
   primaryOffice?: UserToPrimaryOfficeResolver<TParent>
-  catchmentArea?: UserToCatchmentAreaResolver<TParent>
   localRegistrar?: UserToLocalRegistrarResolver<TParent>
   identifier?: UserToIdentifierResolver<TParent>
   signature?: UserToSignatureResolver<TParent>
@@ -5117,15 +5125,6 @@ export interface UserToUnderInvestigationResolver<
 }
 
 export interface UserToPrimaryOfficeResolver<TParent = any, TResult = any> {
-  (
-    parent: TParent,
-    args: {},
-    context: Context,
-    info: GraphQLResolveInfo
-  ): TResult
-}
-
-export interface UserToCatchmentAreaResolver<TParent = any, TResult = any> {
   (
     parent: TParent,
     args: {},