+++ packages/gateway/src/utils/testUtils.ts
@@ -1812,12 +1812,6 @@ export const mockUserDetails = {
   status: 'active',
   practitionerId: '2d11389d-f58e-4d47-a562-b934f1b85936',
   primaryOfficeId: '79776844-b606-40e9-8358-7d82147f702a',
-  catchmentAreaIds: [
-    'b21ce04e-7ccd-4d65-929f-453bc193a736',
-    '95754572-ab6f-407b-b51a-1636cb3d0683',
-    '7719942b-16a7-474a-8af1-cd0c94c730d2',
-    '43ac3486-7df1-4bd9-9b5e-728054ccd6ba'
-  ],
   securityQuestionAnswers: [
     {
       questionKey: 'BIRTH_TOWN',